requirements.txt
flask==3.0.2
opencv-python-headless==4.9.0.80
twilio==8.12.0
numpy==1.26.4
torch==2.2.0
torchvision==0.17.0
pillow==10.2.0
werkzeug==3.0.1
ultralytics==8.1.28
python-dotenv==1.0.1
gunicorn==21.2.0 